Use message from "error" step in title
In status checks, if the run failed because of an "error" pipeline step,
then use the first line of the error message in the title of ChecksOutput,
instead of "error in 'error' step".

However, keep the original behaviour if the error message does not
match the "message" argument of the step, if the step has other
arguments, or if the first line of the message is empty.

Update the test so that it keeps passing,