Only return names in getStageOrBranchName if it's a stage or branch
If `stageOrBranchName` is present, we skip processing errors. This is on
the assumption that if `stageOrBranchName` is present it is a stage or
branch. However, normal steps like `sh` can have labels too and was
returned as the `stageOrBranchName`. This causes labeled steps to not
have their errors processed.

We fix this by returning empty in `getStageOrBranchName` if the node is
not a `BlockStartNode`.